split contextmenu in own command
to do this the userinfo module was created
