+++ src/commands/general.rs
@@ -1,5 +1,6 @@
 use crate::{
-    traits::readable::Readable, types::data::Data, utils::emoji_cache::EmojiCache, Context, Error,
+    traits::context_ext::ContextExt, traits::readable::Readable, types::data::Data,
+    utils::emoji_cache::EmojiCache, utils::userinfo::userinfo_fn, Context, Error,
 };
 use futures::StreamExt;
 use poise::{
@@ -51,78 +52,44 @@ pub async fn ping(ctx: Context<'_>) -> Result<(), Error> {
 
     Ok(())
 }
+
+#[poise::command(context_menu_command = "User info", category = "General", hide_in_help)]
+pub async fn userinfo_ctxmenu(ctx: Context<'_>, user: User) -> Result<(), Error> {
+    ctx.defer_ephemeral().await?;
+
+    let reply = userinfo_fn(ctx, user).await?;
+    if reply.is_none() {
+        return Ok(());
+    }
+    let reply = reply.unwrap();
+    ctx.send(|m| {
+        m.clone_from(&reply);
+        m
+    })
+    .await?;
+    Ok(())
+}
+
 /// User info for a user
 ///
 /// Can be used to get the user info of the specified user. A context menu command is also available
 /// ``userinfo [user]``
-#[poise::command(
-    prefix_command,
-    slash_command,
-    context_menu_command = "User info",
-    category = "General"
-)]
-pub async fn userinfo(ctx: Context<'_>, #[description = "User"] user: User) -> Result<(), Error> {
+#[poise::command(prefix_command, slash_command, category = "General")]
+pub async fn userinfo(
+    ctx: Context<'_>,
+    #[description = "User"] user: Option<User>,
+) -> Result<(), Error> {
     ctx.defer_ephemeral().await?;
-    let color = ctx.data().colors.user_server_info().await;
-
-    let (nickname, joined_at, roles) = match ctx.guild() {
-        Some(guild) => {
-            match guild.member(ctx.discord(), user.id).await {
-                Ok(member) => {
-                    let nick = member.nick.clone().unwrap_or("None".to_string());
-                    let joined_at = match member.joined_at {
-                        Some(joined_at) => joined_at.readable(),
-                        None => "N/A".to_string(),
-                    };
-                    let mut roles = match member.roles(ctx.discord()) {
-                        Some(roles) => roles
-                            .iter()
-                            .map(|role| format!("<@&{}>, ", role.id))
-                            .collect::<String>(),
-                        None => "None".to_string(),
-                    };
-                    // Remove trailing comma and space
-                    roles.pop();
-                    roles.pop();
-
-                    // Make sure it isn't empty
-                    if roles == "" {
-                        roles = "None".to_string()
-                    }
-                    (nick, joined_at, roles)
-                }
-                Err(_) => ("N/A".to_string(), "N/A".to_string(), "N/A".to_string()),
-            }
-        }
-        None => ("N/A".to_string(), "N/A".to_string(), "N/A".to_string()),
-    };
-
-    let mut easter_egg_fields = Vec::new();
-    if ctx.framework().bot_id.0 == user.id.0 {
-        let data = sqlx::query!(r#"SELECT field_name, field_value FROM ttc_easter_egg_botinfo"#)
-            .fetch_all(&*ctx.data().pool)
-            .await?;
-        for row in data {
-            easter_egg_fields.push((row.field_name, row.field_value, false));
-        }
+    let reply = userinfo_fn(ctx, user.unwrap_or(ctx.author().clone())).await?;
+    if reply.is_none() {
+        return Ok(());
     }
-
+    let reply = reply.unwrap();
     ctx.send(|m| {
-        m.embed(|e| {
-            e.author(|a| a.name(user.tag()).icon_url(user.face()))
-                .field("User ID", user.id.0, true)
-                .field("Nickname", nickname, true)
-                .field("Created At", user.id.created_at().readable(), false)
-                .field("Joined At", joined_at, false)
-                .field("Roles", roles, false)
-                .field("Icon URL", user.face(), false)
-                .fields(easter_egg_fields)
-                .color(color)
-        })
-        .ephemeral(true)
+        m.clone_from(&reply);
+        m
     })
     .await?;
-
     Ok(())
 }
